+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/personwithsignificantcontrol/NatureOfControl.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.psc;
+package uk.gov.companieshouse.limitedpartnershipsapi.model.personwithsignificantcontrol;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonEnumDefaultValue;
