+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/personwithsignificantcontrol/dto/PersonWithSignificantControlSubmissionCreatedResponseDto.java
@@ -1,0 +1,4 @@
+package uk.gov.companieshouse.limitedpartnershipsapi.model.personwithsignificantcontrol.dto;
+
+public record PersonWithSignificantControlSubmissionCreatedResponseDto(String id) {
+}
