--- src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/psc/dto/PscSubmissionCreatedResponseDto.java
@@ -1,4 +1,0 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.psc.dto;
-
-public record PscSubmissionCreatedResponseDto(String id) {
-}
