+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/repository/PersonWithSignificantControlRepository.java
@@ -1,0 +1,7 @@
+package uk.gov.companieshouse.limitedpartnershipsapi.repository;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.personwithsignificantcontrol.dao.PersonWithSignificantControlDao;
+
+public interface PersonWithSignificantControlRepository extends MongoRepository<PersonWithSignificantControlDao, String> {
+}
