--- src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/repository/PscRepository.java
@@ -1,7 +1,0 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.repository;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import uk.gov.companieshouse.limitedpartnershipsapi.model.psc.dao.PscDao;
-
-public interface PscRepository extends MongoRepository<PscDao, String> {
-}
