+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/utils/Constants.java
@@ -17,7 +17,7 @@ public class Constants {
     public static final String URL_PARAM_INCORPORATION_ID = "incorporationId";
     public static final String URL_PARAM_GENERAL_PARTNER_ID = "generalPartnerId";
     public static final String URL_PARAM_LIMITED_PARTNER_ID = "limitedPartnerId";
-    public static final String URL_PARAM_PSC_ID = "personsWithSignificantControlId";
+    public static final String URL_PARAM_PERSON_WITH_SIGNIFICANT_CONTROL_ID = "personWithSignificantControlId";
     public static final String TRANSACTION_KEY = "transaction";
     public static final String TRANSACTION_ID_KEY = "transaction_id";
 
@@ -35,13 +35,13 @@ public class Constants {
     public static final String URL_GET_INCORPORATION = "/transactions/%s/incorporation/limited-partnership/%s";
     public static final String URL_GET_LIMITED_PARTNER = "/transactions/%s/limited-partnership/limited-partner/%s";
     public static final String URL_GET_GENERAL_PARTNER = "/transactions/%s/limited-partnership/general-partner/%s";
-    public static final String URL_GET_PSC = "/transactions/%s/limited-partnership/person-with-significant-control/%s";
+    public static final String URL_GET_PERSON_WITH_SIGNIFICANT_CONTROL = "/transactions/%s/limited-partnership/person-with-significant-control/%s";
 
     // Filings
     public static final String FILING_KIND_LIMITED_PARTNERSHIP = "limited-partnership";
     public static final String FILING_KIND_LIMITED_PARTNER = "limited-partnership#limited-partner";
     public static final String FILING_KIND_GENERAL_PARTNER = "limited-partnership#general-partner";
-    public static final String FILING_KIND_PSC = "limited-partnership#person-with-significant-control";
+    public static final String FILING_KIND_PERSON_WITH_SIGNIFICANT_CONTROL = "limited-partnership#person-with-significant-control";
 
     // Payment
     public static final String FILING_PAYMENT_REFERENCE = "payment_reference";
